+++ src/config.ts
@@ -14,6 +14,7 @@
 
 import * as fs from 'fs';
 import * as path from 'path';
+import * as bowerConfig from 'bower-config';
 import {ServerOptions} from './start_server';
 
 function readConfigSync(filename: string, root?: string): any {
@@ -49,10 +50,8 @@ export function getPackageName(options: ServerOptions) {
 }
 
 export function getComponentDir(options: ServerOptions) {
-  // TODO(usergenic): The current behavior of polyserve is to use
-  // bower_components for directory and components for url. We should
-  // honor the value of `directory` of `.bowerrc` if found in the root dir
-  // of the app, to get user-defined defaults.
+  const root = options.root || process.cwd();
+  const bowerDir = bowerConfig.read(root).directory;
   return options.componentDir ||
-      (options.npm ? 'node_modules' : 'bower_components');
+      (options.npm ? 'node_modules' : bowerDir);
 }